webinar/requirements.txt
adaptivecardbuilder==0.0.9
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
frozenlist==1.4.1
future==1.0.0
humanfriendly==10.0
idna==3.6
multidict==6.0.5
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.2.1
git+https://github.com/fbradyirl/webex_bot.git@cc5cd1e20f7cfeb7e0641c378d79f08d7c48572a
webexteamssdk==1.6.1
websockets==11.0.3
yarl==1.9.4
